Workspaces code refactoring
