+++ src/apps/applets/crud/__init__.py
@@ -1,3 +1,3 @@
 from apps.applets.crud.applets import *  # noqa: F401, F403
 from apps.applets.crud.applets_history import *  # noqa: F401, F403
-from apps.applets.crud.roles import *  # noqa: F401, F403
+from apps.workspaces.crud.roles import *  # noqa: F401, F403
