+++ src/apps/applets/db/schemas/__init__.py
@@ -1,2 +1,2 @@
 from apps.applets.db.schemas.applet import *  # noqa: F401, F403
-from apps.applets.db.schemas.user_applet_access import *  # noqa: F401, F403
+from apps.workspaces.db.schemas.user_applet_access import *  # noqa: F401, F403
