+++ src/apps/applets/domain/__init__.py
@@ -10,6 +10,6 @@ from apps.applets.domain.applet_history import (  # noqa: F401, F403
     PublicHistory,
 )
 from apps.applets.domain.applet_link import *  # noqa: F401, F403
-from apps.applets.domain.constants import *  # noqa: F401, F403
 from apps.applets.domain.history import *  # noqa: F401, F403
-from apps.applets.domain.user_applet_access import *  # noqa: F401, F403
+from apps.workspaces.domain.constants import *  # noqa: F401, F403
+from apps.workspaces.domain.user_applet_access import *  # noqa: F401, F403
