+++ src/apps/applets/errors.py
@@ -4,8 +4,6 @@ __all__ = [
     "AppletsError",
     "AppletAlreadyExist",
     "AppletNotFoundError",
-    "UserAppletAccessesNotFound",
-    "AppletAccessDenied",
     "AppletsFolderAccessDenied",
     "AppletLinkNotFoundError",
     "AppletLinkAlreadyExist",
@@ -25,16 +23,6 @@ class AppletAlreadyExist(ValidationError):
 class AppletNotFoundError(NotFoundError):
     def __init__(self, *_, key: str, value: str) -> None:
         super().__init__(message=f"No such applets with {key}={value}.")
-
-
-class UserAppletAccessesNotFound(BaseError):
-    def __init__(self, *_, id_: int) -> None:
-        super().__init__(message=f"No such UserAppletAccess with id={id_}.")
-
-
-class AppletAccessDenied(ValidationError):
-    def __init__(self, *_, message="Access denied to applet.") -> None:
-        super().__init__(message=message)
 
 
 class AppletsFolderAccessDenied(ValidationError):
