+++ src/apps/applets/service/__init__.py
@@ -1,3 +1,3 @@
 from apps.applets.service.applet_history_service import *  # noqa: F401, F403
 from apps.applets.service.applet_service import *  # noqa: F401, F403
-from apps.applets.service.user_applet_access import *  # noqa: F401, F403
+from apps.workspaces.service.user_applet_access import *  # noqa: F401, F403
