+++ src/apps/workspaces/domain/constants.py
